Fix crash on startup when XDG_RUNTIME_DIR is not set.
In Qt 5.8, a message is logged when QStandardPaths::writableLocation() is
called for the first time, telling you that XDG_RUNTIME_DIR is not set.
QuiteRSS tries to log this to the dataDir, but we called
QStandardPaths::writeableLocation() while *initializing* the dataDir.
So we try to return an uninitialized QString member, and crash horribly.

This is a crude fix, that tracks whether the dataDir is initialized yet,
and if it's not, just skip logging messages entirely. (It might be nicer
to print them to stderr, but that might also be ugly... maybe we should
do that, but only in debugging mode? I'm not sure.)